+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - compatibility fix
+- nerfed current branch name
 
 ### Changed